Filter Problem on user_management solved. Added Show columns and Show column method on users section.

+++ libs/footer.php
@@ -13,6 +13,7 @@
 
     <script>
     
+    //Area de usuarios
     $('#userFilterForm').submit(function(e)
     {
         e.preventDefault();
@@ -24,44 +25,55 @@
     $("#btnListar").click(function()
     {
       $.post('libs/user_management.php?action=getUsers&filter='+$('#filtrarPor option:selected').val()
-      +'&order='+$('#ordenarPor option:selected').val(),
+      +'&value='+$('#filtro').val()+'&order='+$('#ordenarPor option:selected').val(),
       function(response){
+
         showColumns(response);
       });
     });
 
     function showColumns(data)
     {
-        console.log(data);
-        var users = JSON.parse(data);
-        if(users.length > 1)
+        var users = JSON.parse(data);   console.log(users);
+        tBody.empty();
+        if(Array.isArray(users) && users.length > 0)
         {
           users.forEach(function(element)
           {
             showColumn(element);
           });
         }
-        else if(users.length == 1)
+        else if(typeof(users) == 'object' && users != null)
         {
           showColumn(users);
-        }        
+        }      
     }
     function showColumn(element)
     {
-      tBody.append('<tr>');
-      tBody.append('<td>'+element.nombre+'</td>');
-      tBody.append('<td>'+element.apellido+'</td>');
-      tBody.append('<td>'+element.direccion+'</td>');
-      tBody.append('<td>'+element.telefono+'</td>');
-      tBody.append('<td>'+element.email+'</td>');
-      tBody.append('<td>'+element.username+'</td>');
-      tBody.append('<td>'+element.created_at+'</td>');
-      tBody.append('<td><a href="libs/user_management.php?action=getUser&id='+element.id+'">Detalle</a></td>');
-      tBody.append('<td><a href="libs/user_management.php?action=updateUser&id='+element.id+'">Modificar</a></td>');
-      tBody.append('<td><a href="libs/user_management.php?action=deleteUsers&id='+element.id+'">Eliminar</a></td>');
-      tBody.append('</tr><a></a>');
+      if(element != null)
+      {
+        tBody.append('<tr>');
+        tBody.append('<td>'+element.nombre+'</td>');
+        tBody.append('<td>'+element.apellido+'</td>');
+        tBody.append('<td>'+element.direccion+'</td>');
+        tBody.append('<td>'+element.telefono+'</td>');
+        tBody.append('<td>'+element.email+'</td>');
+        tBody.append('<td>'+element.username+'</td>');
+        tBody.append('<td>'+element.created_at+'</td>');
+        tBody.append('<td><a href="libs/user_management.php?action=getUser&id='+element.id+'">Detalle</a></td>');
+        tBody.append('<td><a href="libs/user_management.php?action=updateUser&id='+element.id+'">Modificar</a></td>');
+        tBody.append('<td><a href="libs/user_management.php?action=deleteUsers&id='+element.id+'">Eliminar</a></td>');
+        tBody.append('</tr><a></a>');
+      }
     }
 
+    //Fin Area de usuarios
+
+
+    //Area de clientes.
+
+
+    //Fin Area de clientes.
 
     </script>
   </body>
